requirements.txt
flask
openai==0.28
pydub
moviepy
numpy
PyPDF2
python-docx